refactor: split budget API module by responsibility
