+++ frontend/src/api/budgets/cache.ts
@@ -1,0 +1,10 @@
+import type { QueryClient } from '@tanstack/react-query';
+import { invalidateBudgets, invalidateDashboardBudgets } from '@/api/cacheInvalidation';
+
+/**
+ * Invalidates budget views after budget mutations change dashboard rollups
+ */
+export function invalidateBudgetActivity(queryClient: QueryClient) {
+  invalidateBudgets(queryClient);
+  invalidateDashboardBudgets(queryClient);
+}
